[Feat]#5-NetworkManager 만들기
