+++ BookShelf/View/BookDetailViewController.swift
@@ -10,13 +10,13 @@ import UIKit
 class BookDetailViewController: UIViewController {
   
   var document: Document?
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-      view.backgroundColor = .white
-    }
+  
+  override func viewDidLoad() {
+    super.viewDidLoad()
+    view.backgroundColor = .white
+  }
   
   
   
-
+  
 }
